RavenDB-22218 - Addressed PR Comments
